Fix FileResponse doing blocking I/O in the event loop
The call to os.stat was run in the executor but the
call to `is_file()` was not. Additionally this resulted
in two syscalls to `stat` since both `is_file` and
`stat` will call `stat` under the hood. The code
has been refactored to ensure there is only one
`stat` syscall and it happens in the executor